(1248) Split up changelogModel
